Refactor Account module classes with improvements and added unit tests

+++ account/src/main/java/eu/voops/account/AccountService.java
@@ -1,9 +1,12 @@
 package eu.voops.account;
 
+import eu.voops.account.exception.ProfileExistException;
 import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
+import lombok.NonNull;
 import org.springframework.stereotype.Service;
 
+import java.util.NoSuchElementException;
 import java.util.Random;
 import java.util.stream.Collectors;
 
@@ -14,13 +17,18 @@ public class AccountService {
     private AccountRepository repository;
 
     public void createAccount(Account account) {
-        repository.save(account);
+        if (repository.existsByInternalId(account.getInternalId())) {
+            throw new ProfileExistException("Profile Exist");
+        } else {
+            repository.save(account);
+        }
     }
 
+    @Transactional
     public String makeAccountNumber() {
         Random random = new Random(System.currentTimeMillis());
-        String account;
         final int accountSize = 12;
+        String account;
 
         do {
             account = random.ints(0, 10)
@@ -28,12 +36,16 @@ public class AccountService {
                     .limit(accountSize)
                     .collect(Collectors.joining());
         } while (repository.existsByAccountNumber(account));
-        
+
         return account;
     }
 
     @Transactional
     public void emergencyDelete(String internalId) {
-        repository.deleteByInternalId(internalId);
+        if (repository.existsByInternalId(internalId)) {
+            repository.deleteByInternalId(internalId);
+        } else {
+            throw new NoSuchElementException("Profile does not exist");
+        }
     }
 }
